Feature/route state

+++ app/src/main/java/com/foodpanda/urbanninja/manager/ApiManager.java
@@ -217,7 +217,7 @@ public class ApiManager implements Managable {
         });
     }
 
-    public void notifyActionPerformed(int routeId, Action action) {
+    public void notifyActionPerformed(long routeId, Action action) {
         PerformActionWrapper performActionWrapper = new PerformActionWrapper(action, new DateTime());
         Call<Stop> call = service.notifyActionPerformed(routeId, performActionWrapper);
         call.enqueue(new RetryActionCallback<>(call, routeId, performActionWrapper));
